+++ CHANGELOG.md
@@ -6,8 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.3.5] - 2025-06-13
 ### Fixed
 - Empty trade files in trade generator in 1.21.1 & 1.21.4
+- Config saving in modmenu
 
 ## [4.3.4] - 2025-05-23
 ### Changed